refactor: domain 패키지로 도메인 이동
