+++ src/main/java/subway/controller/LineController.java
@@ -5,6 +5,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import subway.domain.Line;
+import subway.domain.Section;
+import subway.domain.Station;
 import subway.dto.*;
 import subway.exception.DuplicateLineNameException;
 import subway.exception.LineNotFoundException;
