+++ src/main/java/subway/controller/SectionController.java
@@ -4,10 +4,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.dto.Line;
-import subway.dto.Section;
+import subway.domain.Line;
+import subway.domain.Section;
 import subway.dto.SectionRequest;
-import subway.dto.SectionResponse;
 import subway.exception.InvalidSectionInsertException;
 import subway.exception.NotEnoughLengthToDeleteSectionException;
 import subway.exception.StationNotFoundException;
