+++ src/main/java/subway/dao/LineDao.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
-import subway.dto.Line;
+import subway.domain.Line;
 
 import java.util.List;
