+++ src/main/java/subway/dao/StationDao.java
@@ -4,7 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
-import subway.dto.Station;
+import subway.domain.Station;
 
 import java.util.List;
