+++ src/main/java/subway/domain/Section.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.domain;
 
 public class Section {
