+++ src/main/java/subway/domain/Sections.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.domain;
 
 import java.util.HashMap;
 import java.util.List;
