+++ src/main/java/subway/dto/StationResponse.java
@@ -1,5 +1,7 @@
 package subway.dto;
 
+import subway.domain.Station;
+
 public class StationResponse {
     private Long id;
     private String name;
