+++ src/main/java/subway/service/LineService.java
@@ -3,7 +3,7 @@ package subway.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import subway.dao.LineDao;
-import subway.dto.Line;
+import subway.domain.Line;
 import subway.exception.DuplicateLineNameException;
 import subway.exception.LineNotFoundException;
